Cygwin fixes.
Patch submitted by Stig Petter Olsrød <stigpo@users.sourceforge.net>.

+++ flthdr.c
@@ -60,13 +60,14 @@ process_file(char *ifile, char *ofile)
 {
 	int old_flags, old_stack, new_flags, new_stack;
 	FILE *ifp, *ofp;
+	int ofp_is_pipe = 0;
 	struct flat_hdr old_hdr, new_hdr;
 	char tfile[256];
 	char tfile2[256];
 
 	*tfile = *tfile2 = '\0';
 
-	if ((ifp = fopen(ifile, "r")) == NULL) {
+	if ((ifp = fopen(ifile, "rb")) == NULL) {
 		fprintf(stderr, "Cannot open %s\n", ifile);
 		return;
 	}
@@ -189,7 +190,7 @@ process_file(char *ifile, char *ofile)
 
 	strcpy(tfile, "/tmp/flatXXXXXX");
 	mkstemp(tfile);
-	if ((ofp = fopen(tfile, "w")) == NULL) {
+	if ((ofp = fopen(tfile, "wb")) == NULL) {
 		fprintf(stderr, "Failed to open %s for writing\n", tfile);
 		unlink(tfile);
 		unlink(tfile2);
@@ -214,7 +215,7 @@ process_file(char *ifile, char *ofile)
 		mkstemp(tfile2);
 		
 		if (old_flags & FLAT_FLAG_GZDATA) {
-			tfp = fopen(tfile2, "w");
+			tfp = fopen(tfile2, "wb");
 			if (!tfp) {
 				fprintf(stderr, "Failed to open %s for writing\n", tfile2);
 				exit(1);
@@ -225,12 +226,16 @@ process_file(char *ifile, char *ofile)
 		}
 
 		sprintf(cmd, "gunzip >> %s", tfile2);
-		tfp = popen(cmd, "w");
+		tfp = popen(cmd, "wb");
+		if(!tfp) {
+			perror("popen");
+			exit(1);
+		}
 		transferr(ifp, tfp, -1);
-		fclose(tfp);
+		pclose(tfp);
 
 		fclose(ifp);
-		ifp = fopen(tfile2, "r");
+		ifp = fopen(tfile2, "rb");
 		if (!ifp) {
 			fprintf(stderr, "Failed to open %s for reading\n", tfile2);
 			unlink(tfile);
@@ -243,14 +248,16 @@ process_file(char *ifile, char *ofile)
 		printf("zflat %s --> %s\n", ifile, ofile);
 		fclose(ofp);
 		sprintf(cmd, "gzip -9 -f >> %s", tfile);
-		ofp = popen(cmd, "w");
+		ofp = popen(cmd, "wb");
+		ofp_is_pipe = 1;
 	} else if (new_flags & FLAT_FLAG_GZDATA) {
 		printf("zflat-data %s --> %s\n", ifile, ofile);
 		transferr(ifp, ofp, ntohl(new_hdr.data_start) -
 				sizeof(struct flat_hdr));
 		fclose(ofp);
 		sprintf(cmd, "gzip -9 -f >> %s", tfile);
-		ofp = popen(cmd, "w");
+		ofp = popen(cmd, "wb");
+		ofp_is_pipe = 1;
 	}
 
 	if (!ofp) { /* can only happen if using gzip/gunzip */
@@ -271,7 +278,10 @@ process_file(char *ifile, char *ofile)
 	}
 
 	fclose(ifp);
-	fclose(ofp);
+	if (ofp_is_pipe)
+		pclose(ofp);
+	else
+		fclose(ofp);
 
 	/* cheat a little here to preserve file permissions */
 	sprintf(cmd, "cp %s %s", tfile, ofile);
